pyproject.toml
[project]
name = "healthcare-ml-project"
version = "1.0.0"
description = "Healthcare patient test-result prediction API with weekly XGBoost retraining"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn==0.32.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "sqlalchemy==2.0.36",
    "psycopg2-binary==2.9.10",
    "alembic==1.13.3",
    "scikit-learn==1.5.2",
    "xgboost==2.1.3",
    "pandas==2.1.4",
    "numpy==1.26.4",
    "joblib==1.4.2",
    "kaggle==1.6.17",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "httpx==0.27.2",
]

[tool.uv]
dev-dependencies = [
    "pytest==8.3.3",
    "httpx==0.27.2",
]
